intel-feed/requirements.txt
feedparser==6.0.11
anthropic==0.42.0
requests==2.31.0
pyyaml==6.0.1
supabase==2.4.0
